ALMOST FINISHEDgit add *git add *
