+++ src/exec/heredoc.c
@@ -6,54 +6,39 @@
 /*   By: emonacho <emonacho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 13:25:02 by emonacho          #+#    #+#             */
-/*   Updated: 2025/06/25 14:13:59 by emonacho         ###   ########.fr       */
+/*   Updated: 2025/06/25 16:08:56 by emonacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
 
-int	put_in_heredoc(char *line, int fd)
+int	heredoc_loop(t_shell *s, t_ast *node)
 {
-	char	*new_line;
+	int	i;
 
-	new_line = ft_strjoin(line, "\n");
-	if (!new_line)
-		return (1);
-	ft_putstr_fd(new_line, fd);
-	free(new_line);
-	return (0);
-}
-
-int	is_delimiter(char *line, char *delimiter)
-{
-	size_t	i;
-	size_t	j;
-
-	i = 0;
-	j = 0;
-	while (delimiter[j] && (delimiter[j] == '\'' || delimiter[j] == '\"'))
-		j++;
-	if (line[i] == delimiter[j])
+	s->heredoc_fd = 0;
+	i = -1;
+	while (++i < node->data.s_exec.heredoc_count)
 	{
-		while (line[i] && delimiter[j]
-			&& (line[i] == delimiter[j]))
+		if (s->heredoc_fd > 0)
+			if (close(s->heredoc_fd) < 0)
+				return (-1);
+		s->fd = 0;
+		if (i == node->data.s_exec.heredoc_count - 1)
 		{
-			i++;
-			j++;
+			if (write_heredoc(s, node->data.s_exec.heredoc_list[i], 1) != 0)
+				return (-1);
 		}
-		if (line[i] == '\0' && (delimiter[j] == '\0'
-				|| delimiter[j] == '\'' || delimiter[j] == '\"'))
-			return (1);
+		else
+		{
+			if (write_heredoc(s, node->data.s_exec.heredoc_list[i], 0) != 0)
+				return (-1);
+		}
+		s->heredoc_fd = redir_in(s, s->heredoc_tmp, 0);
+		if (s->heredoc_fd < 0)
+			return (-1);
 	}
 	return (0);
-}
-
-int	open_heredoc(t_shell *s)
-{
-	s->fd = open(s->heredoc_tmp, O_WRONLY | O_CREAT | O_TRUNC, 0600);
-	if (s->fd < 0)
-		terminate_shell(s);
-	return (s->fd);
 }
 
 int	write_heredoc(t_shell *s, char *del, int to_expand)
@@ -63,17 +48,17 @@ int	write_heredoc(t_shell *s, char *del, int to_expand)
 	reset_prompt(s, HEREDOC_PROMPT);
 	while (1)
 	{
+		if (g_sig != 0)
+			break ;
 		s->line = readline("> ");
 		if (!s->line)
 			break ;
 		if (is_delimiter(s->line, del))
 			break ;
 		if (to_expand == 1)
-		{
 			if ((del[0] != '\'' && del[ft_strlen(del) - 1] != '\'')
 				&& (del[0] != '\"' && del[ft_strlen(del) - 1] != '\"'))
 				expand(s->numerr, s->env_list, &(s->line));
-		}
 		if (put_in_heredoc(s->line, s->fd) != 0)
 			return (-1);
 		w_free((void **)&s->line);
@@ -88,9 +73,8 @@ int	handle_heredoc(t_shell *s, t_ast *node)
 {
 	pid_t	heredoc_pid;
 
-	g_sig = 0;
-	s->child_exit = 1;
-	setup_signals(s, HEREDOC_SIGNALS);
+	if (handle_termios(s, 0) != 0)
+		return (-1);
 	signal(SIGINT, SIG_IGN);
 	signal(SIGQUIT, SIG_IGN);
 	heredoc_pid = fork();
@@ -98,14 +82,31 @@ int	handle_heredoc(t_shell *s, t_ast *node)
 		return (print_error(&s->numerr, EPIPE));
 	if (heredoc_pid == 0)
 	{
-		signal(SIGINT, SIG_DFL);
+		setup_signals(s, HEREDOC_SIGNALS);
 		if (heredoc_loop(s, node) == -1)
 			kill_children(s);
 		clean_free(s);
-		exit(0);
+		if (g_sig == SIGINT)
+			exit(130);
+		exit(s->numerr);
 	}
 	waitheredoc(&s->numerr, heredoc_pid);
-	setup_signals(s, DEFAULT_SIGNALS);
+	if (handle_termios(s, 1) != 0)
+		return (-1);
 	s->heredoc_fd = redir_in(s, s->heredoc_tmp, 0);
+	setup_signals(s, DEFAULT_SIGNALS);
 	return (s->heredoc_fd);
+}
+
+void	waitheredoc(uint8_t *numerr, pid_t pid)
+{
+	int	status;
+
+	waitpid(pid, &status, 0);
+	if (WIFSIGNALED(status))
+		*numerr = 128 + WTERMSIG(status);
+	else if (WIFEXITED(status))
+		*numerr = WEXITSTATUS(status);
+	else
+		*numerr = 1;
 }
